@@ -1,7 +1,8 @@
 import { Router, ActivatedRoute } from '@angular/router';
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 
 import { Item } from '../../item.model';
+import {ItemService} from '../../item.service';
 
 @Component({
   selector: 'app-item-list-element',
@@ -11,16 +12,17 @@ import { Item } from '../../item.model';
 export class ItemListElementComponent implements OnInit {
   @Input() item: Item;
   @Input() index: number;
+  private checked = false;
   private clicked = false;
-
+  private picIsOpen = false;
   constructor(private route: ActivatedRoute,
-              private router: Router) { }
+              private router: Router,
+              private itemService: ItemService) { }
 
-  ngOnInit() {
-  }
+  ngOnInit() { }
 
   onClick() {
-    if (!this.clicked) {
+    if (!this.clicked && (this.item.desc || this.item.attribute )) {
       this.clicked = true;
     } else {
       const calling_id = this.item.id ? this.item.id : this.index;
@@ -31,4 +33,17 @@ export class ItemListElementComponent implements OnInit {
   }
   onClose() { this.clicked = false; }
 
+  onToggle() {
+    this.checked = !this.checked;
+    this.itemService.onCheckBoxChange({
+      'id': this.item.id,
+      'checked': this.checked
+    });
+  }
+
+  openPicture() {
+    if (this.item.images.length > 0 ) {
+      this.picIsOpen = !this.picIsOpen;
+    }
+  }
 }
