@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Subject } from 'rxjs';
-import { Observable } from 'rxjs';
+import { Subject ,  Observable } from 'rxjs';
 import * as moment from 'moment';
 
 @Injectable()
